// src/main/features/fca/model/fca.ts
export enum FCA_PATHS {
    /**
    * @ROUTES
    */

     SUPPLIERS = '/fca-tasklist',
     SELECTED_SERVICES = '/fca/services',
     POST_SHORTLISTED_SUPPLIERS = '/fca/shortlisted/suppliers',
     SHORTLISTED_SUPPLIERS = '/fca/shortlisted/suppliers',
     SUPPLIERS_LIST = '/fca/supplier/list',
     SUPPLIER_RATECARD = '/fca/supplier/ratecard',
     FCA_PROCUREMENT='/fca/projects/create-or-choose',
     SHORTLISTED_SUPPLIERS_NEXTSTEP='/fca/shortlisted/suppliers/nextstep',
    //FCA Constant
     ROUTE_TO_CREATE_SUPPLIER='/projects/routeToCreateSupplier',
     CREATE_SUPPLIER_SHORTLIST='/fca/create-supplier-shortlist',
     SELECT_SERVICES='/fca/select-services',
     POST_SELECT_SERVICES='/fca/select-services',
     NEXT_STEP='/fca/next-step',
     EXPORT_SUPPLIER_FILTERING='/fca/supplier-filtering/export',
     GET_NAME_YOUR_PROJECT = '/fca/name-your-project',
     GET_LEAD_PROCUEMENT = '/fca/procurement-lead',
     GET_ADD_COLLABORATOR = '/fca/add-collaborators',
     POST_PROJECT_NAME = '/fca/name',
     POST_ADD_COLLABORATOR = '/fca/get-collaborator-detail',
     POST_ADD_COLLABORATOR_TO_JAGGER = '/fca/add-collaborator-detail',
     PUT_LEAD_PROCUREMENT = '/fca/procurement-lead',
     POST_PROCEED_COLLABORTORS = '/fca/proceed-collaborators',
     POST_ADD_COLLABORATOR_JSENABLED = '/fca/get-collaborator-detail/js-enabled',
     POST_DELETE_COLLABORATOR_TO_JAGGER='/fca/delete-collaborators',
     GET_USER_PROCUREMENT = '/fca/users-procurement-lead',

}